Organize files into include and source directories
