+++ include/NetworkDefinitions.h
@@ -1,0 +1,10 @@
+#pragma once
+
+using ClientID = int;
+using PortNumber = unsigned short;
+
+enum class Network 
+{
+	HighestTimestamp = 2147483647, ClientTimeout = 10000, ServerPort = 5600,
+	NullID = -1, PlayerUpdateDelim = -1
+};
