+++ source/Text.cpp
